\name{BayesX-package}
\alias{BayesX-package}
\alias{BayesX}

\docType{package}

\title{
R Utilities Accompanying the Software Package BayesX
}

\description{
This package provides functionality for exploring and visualising
estimation results obtained with the software package BayesX for
structured additive regression. It also provides functions that allow to
read, write and manipulate map objects that are required in spatial
analyses performed with BayesX. 
}

\details{
\tabular{ll}{
Package: \tab BayesX\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2009-03-10\cr
License: \tab GPL-2\cr
}}

\author{Thomas Kneib, Felix Heinzl, Andreas Brezger, Daniel Sabanes Bove

Maintainer: Thomas Kneib <thomas.kneib@stat.uni-muenchen.de>}

\references{\url{http://www.stat.uni-muenchen.de/~bayesx/}}

\keyword{package}
