\name{BayesX-package}
\alias{BayesX-package}
\docType{package}

\title{
  R Utilities Accompanying the Software Package BayesX
}

\description{
  This package provides functionality for exploring and visualising
  estimation results obtained with the software package \pkg{BayesX}
  for structured additive regression. It also provides functions that allow
  to read, write and manipulate map objects that are required in spatial
  analyses performed with \pkg{BayesX}. 
}

\author{Nikolaus Umlauf, Thomas Kneib, Nadja Klein, Felix Heinzl, Andreas Brezger, Daniel Sabanes Bove}

\references{
  Belitz C, Brezger A, Kneib T, Lang S (2011). \pkg{BayesX} - Software for Bayesian Inference in 
    Structured Additive Regression Models. Version 2.0.1. 
    URL \url{http://www.BayesX.org}.
}

\keyword{package}
